test: consolidates list view tests
