+++ test/fields/collections/Text/e2e.spec.ts
@@ -3,12 +3,9 @@ import type { GeneratedTypes } from 'helpers/sdk/types.js'
 
 import { expect, test } from '@playwright/test'
 import { openListColumns } from 'helpers/e2e/openListColumns.js'
-import { openListFilters } from 'helpers/e2e/openListFilters.js'
 import { toggleColumn } from 'helpers/e2e/toggleColumn.js'
 import { upsertPrefs } from 'helpers/e2e/upsertPrefs.js'
 import path from 'path'
-import { wait } from 'payload/shared'
-import * as qs from 'qs-esm'
 import { fileURLToPath } from 'url'
 
 import type { PayloadTestSDK } from '../../../helpers/sdk/index.js'
@@ -167,84 +164,6 @@ describe('Text', () => {
     await expect(textCell).toHaveText(textDoc.text)
   })
 
-  test('should hide field in column selector when admin.disableListColumn', async () => {
-    await page.goto(url.list)
-    await page.locator('.list-controls__toggle-columns').click()
-
-    await expect(page.locator('.column-selector')).toBeVisible()
-
-    // Check if "Disable List Column Text" is not present in the column options
-    await expect(
-      page.locator(`.column-selector .column-selector__column`, {
-        hasText: exactText('Disable List Column Text'),
-      }),
-    ).toBeHidden()
-  })
-
-  test('should show field in filter when admin.disableListColumn is true', async () => {
-    await page.goto(url.list)
-    await openListFilters(page, {})
-    await page.locator('.where-builder__add-first-filter').click()
-
-    const initialField = page.locator('.condition__field')
-    await initialField.click()
-
-    await expect(
-      initialField.locator(`.rs__menu-list:has-text("Disable List Column Text")`),
-    ).toBeVisible()
-  })
-
-  test('should display field in list view column selector despite admin.disableListFilter', async () => {
-    await page.goto(url.list)
-    await page.locator('.list-controls__toggle-columns').click()
-
-    await expect(page.locator('.column-selector')).toBeVisible()
-
-    // Check if "Disable List Filter Text" is present in the column options
-    await expect(
-      page.locator(`.column-selector .column-selector__column`, {
-        hasText: exactText('Disable List Filter Text'),
-      }),
-    ).toBeVisible()
-  })
-
-  test('should disable field when admin.disableListFilter is true but still exists in the query', async () => {
-    await page.goto(
-      `${url.list}${qs.stringify(
-        {
-          where: {
-            or: [
-              {
-                and: [
-                  {
-                    disableListFilterText: {
-                      equals: 'Disable List Filter Text',
-                    },
-                  },
-                ],
-              },
-            ],
-          },
-        },
-        { addQueryPrefix: true },
-      )}`,
-    )
-
-    await openListFilters(page, {})
-
-    const condition = page.locator('.condition__field')
-    await expect(condition.locator('input.rs__input')).toBeDisabled()
-    await expect(page.locator('.condition__operator input.rs__input')).toBeDisabled()
-    await expect(page.locator('.condition__value input.condition-value-text')).toBeDisabled()
-    await expect(condition.locator('.rs__single-value')).toHaveText('Disable List Filter Text')
-    await page.locator('button.condition__actions-add').click()
-    const condition2 = page.locator('.condition__field').nth(1)
-    await condition2.click()
-    await expect(
-      condition2?.locator('.rs__menu-list:has-text("Disable List Filter Text")'),
-    ).toBeHidden()
-  })
-
   test('should respect admin.disableListColumn despite preferences', async () => {
     await upsertPrefs<Config, GeneratedTypes<any>>({
       payload,
@@ -269,19 +188,6 @@ describe('Text', () => {
 
     await expect(page.locator('#heading-disableListColumnText')).toBeHidden()
     await expect(page.locator('table .row-1 .cell-disableListColumnText')).toBeHidden()
-  })
-
-  test('should hide field in filter when admin.disableListFilter is true', async () => {
-    await page.goto(url.list)
-    await openListFilters(page, {})
-    await page.locator('.where-builder__add-first-filter').click()
-
-    const initialField = page.locator('.condition__field')
-    await initialField.click()
-
-    await expect(
-      initialField.locator(`.rs__option :has-text("Disable List Filter Text")`),
-    ).toBeHidden()
   })
 
   test('should display i18n label in cells when missing field data', async () => {
@@ -332,137 +238,5 @@ describe('Text', () => {
     await saveDocAndAssert(page)
     await expect(field.locator('.rs__value-container')).toContainText(input)
     await expect(field.locator('.rs__value-container')).toContainText(furtherInput)
-  })
-
-  test('should reset filter conditions when adding additional filters', async () => {
-    await page.goto(url.list)
-
-    // open the first filter options
-    await openListFilters(page, {})
-    await page.locator('.where-builder__add-first-filter').click()
-
-    const firstInitialField = page.locator('.condition__field')
-    const firstOperatorField = page.locator('.condition__operator')
-    const firstValueField = page.locator('.condition__value >> input')
-
-    await firstInitialField.click()
-    const firstInitialFieldOptions = firstInitialField.locator('.rs__option')
-    await firstInitialFieldOptions.locator('text=text').first().click()
-    await expect(firstInitialField.locator('.rs__single-value')).toContainText('Text')
-
-    await firstOperatorField.click()
-    await firstOperatorField.locator('.rs__option').locator('text=equals').click()
-
-    await firstValueField.fill('hello')
-
-    await wait(500)
-
-    await expect(firstValueField).toHaveValue('hello')
-
-    // open the second filter options
-    await page.locator('.condition__actions-add').click()
-
-    const secondLi = page.locator('.where-builder__and-filters li:nth-child(2)')
-
-    await expect(secondLi).toBeVisible()
-
-    const secondInitialField = secondLi.locator('.condition__field')
-    const secondOperatorField = secondLi.locator('.condition__operator >> input')
-    const secondValueField = secondLi.locator('.condition__value >> input')
-
-    await expect(secondInitialField.locator('.rs__single-value')).toContainText('Text')
-    await expect(secondOperatorField).toHaveValue('')
-    await expect(secondValueField).toHaveValue('')
-  })
-
-  test('should not re-render page upon typing in a value in the filter value field', async () => {
-    await page.goto(url.list)
-
-    // open the first filter options
-    await openListFilters(page, {})
-    await page.locator('.where-builder__add-first-filter').click()
-
-    const firstInitialField = page.locator('.condition__field')
-    const firstOperatorField = page.locator('.condition__operator')
-    const firstValueField = page.locator('.condition__value >> input')
-
-    await firstInitialField.click()
-    const firstInitialFieldOptions = firstInitialField.locator('.rs__option')
-    await firstInitialFieldOptions.locator('text=text').first().click()
-    await expect(firstInitialField.locator('.rs__single-value')).toContainText('Text')
-
-    await firstOperatorField.click()
-    await firstOperatorField.locator('.rs__option').locator('text=equals').click()
-
-    // Type into the input field instead of filling it
-    await firstValueField.click()
-    await firstValueField.type('hello', { delay: 100 }) // Add a delay to simulate typing speed
-
-    // Wait for a short period to see if the input loses focus
-    await page.waitForTimeout(500)
-
-    // Check if the input still has the correct value
-    await expect(firstValueField).toHaveValue('hello')
-  })
-
-  test('should still show second filter if two filters exist and first filter is removed', async () => {
-    await page.goto(url.list)
-
-    // open the first filter options
-    await openListFilters(page, {})
-    await page.locator('.where-builder__add-first-filter').click()
-
-    const firstInitialField = page.locator('.condition__field')
-    const firstOperatorField = page.locator('.condition__operator')
-    const firstValueField = page.locator('.condition__value >> input')
-
-    await firstInitialField.click()
-    const firstInitialFieldOptions = firstInitialField.locator('.rs__option')
-    await firstInitialFieldOptions.locator('text=text').first().click()
-    await expect(firstInitialField.locator('.rs__single-value')).toContainText('Text')
-
-    await firstOperatorField.click()
-    await firstOperatorField.locator('.rs__option').locator('text=equals').click()
-
-    await firstValueField.fill('hello')
-
-    await wait(500)
-
-    await expect(firstValueField).toHaveValue('hello')
-
-    // open the second filter options
-    await page.locator('.condition__actions-add').click()
-
-    const secondLi = page.locator('.where-builder__and-filters li:nth-child(2)')
-
-    await expect(secondLi).toBeVisible()
-
-    const secondInitialField = secondLi.locator('.condition__field')
-    const secondOperatorField = secondLi.locator('.condition__operator')
-    const secondValueField = secondLi.locator('.condition__value >> input')
-
-    await secondInitialField.click()
-    const secondInitialFieldOptions = secondInitialField.locator('.rs__option')
-    await secondInitialFieldOptions.locator('text=text').first().click()
-    await expect(secondInitialField.locator('.rs__single-value')).toContainText('Text')
-
-    await secondOperatorField.click()
-    await secondOperatorField.locator('.rs__option').locator('text=equals').click()
-
-    await secondValueField.fill('world')
-    await expect(secondValueField).toHaveValue('world')
-
-    await wait(500)
-
-    const firstLi = page.locator('.where-builder__and-filters li:nth-child(1)')
-    const removeButton = firstLi.locator('.condition__actions-remove')
-
-    // remove first filter
-    await removeButton.click()
-
-    const filterListItems = page.locator('.where-builder__and-filters li')
-    await expect(filterListItems).toHaveCount(1)
-
-    await expect(firstValueField).toHaveValue('world')
   })
 })
